rename src/styling/ to src/styles/
