+++ src/components/ui/Buttons.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Pressable, Text } from 'react-native';
 
-const Button = ({ onPress, text, styles }) => {
+const Button = ({ onPress, text, styles, ...props }) => {
     return (
         <Pressable
             style={
@@ -9,13 +9,14 @@ const Button = ({ onPress, text, styles }) => {
                     styles.button]}
             onPress={onPress}
             role='button'
+            {...props}
         >
             <Text style={styles.buttonText}>{text}</Text>
         </Pressable>
     );
 };
 
-const SmallButton = ({ onPress, text, styles }) => {
+const SmallButton = ({ onPress, text, styles, ...props }) => {
     return (
         <Pressable
             style={
@@ -25,10 +26,26 @@ const SmallButton = ({ onPress, text, styles }) => {
             }
             onPress={onPress}
             role='button'
+            {...props}
         >
             <Text style={styles.iconButtonText}>{text}</Text>
         </Pressable>
     );
 };
 
-export { Button, SmallButton };
+const NavButton = ({ onPress, text, styles, ...props }) => {
+    return (
+        <Pressable
+            style={
+                ({ pressed }) => [{ opacity: pressed ? 0.5 : 1.0}
+                ]}
+            onPress={onPress}
+            role='link'
+            {...props}
+        >
+            <Text style={styles.navigationLink}>{text}</Text>
+        </Pressable>
+    );
+};
+
+export { Button, SmallButton, NavButton };
